chore: move redux store files into a store dir
